requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python==1.5.3
Pillow==9.5.0
airflow-provider-duckdb==0.0.2
apache-airflow-providers-snowflake==4.0.5
torch==2.0.0
torchvision==0.15.1
transformers==4.28.1
apache-airflow-providers-slack==7.2.0
evaluate==0.4.0
scikit-learn==1.2.2